+++ packages/ngx-layout/src/lib/sidebar/public-api.ts
@@ -1,4 +1,0 @@
-export { SidebarComponent } from './components/sidebar/sidebar.component';
-export { SidebarItemComponent } from './components/sidebar-item/sidebar-item.component';
-export { Sidebar, SidebarTheme, SidebarState, SidebarItem } from './types/sidebar.types';
-export { SidebarModule } from './sidebar.module';